// web/src/api/generated/models/pmAnalyticsMeasureResponse.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * UniPM.Api | v1
 * OpenAPI spec version: 1.0.0
 */

export interface PmAnalyticsMeasureResponse {
  /** @nullable */
  department: string | null
  /** @pattern ^-?(?:0|[1-9]\d*)$ */
  numerator: number | string
  /** @pattern ^-?(?:0|[1-9]\d*)$ */
  denominator: number | string
  /**
   * @nullable
   * @pattern ^-?(?:0|[1-9]\d*)(?:\.\d+)?$
   */
  value: number | string | null
  unit: string
  isMeasurable: boolean
}