// web/src/api/generated/models/pmAnalyticsSourceResponse.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * UniPM.Api | v1
 * OpenAPI spec version: 1.0.0
 */

export interface PmAnalyticsSourceResponse {
  scheduleId: string
  assetId: string
  /** @nullable */
  inspectionId: string | null
  assetCode: string
  /** @nullable */
  department: string | null
  pmCycle: string
  deadline: string
  /** @nullable */
  inspectionCompletedAt: string | null
  timeliness: string
  condition: string
  /** @nullable */
  formStatus: string | null
}